Add x64 support and harden skin loading

+++ LR2/structure.h
@@ -4,6 +4,7 @@
 #include <vfw.h>
 #include <vector>
 #include <array>
+#include <stdint.h>
 #include "FMOD/fmod.h"
 #include "strclass.h"
 typedef unsigned char   undefined;
@@ -1553,7 +1554,7 @@ struct game {
 	int po4MainMenuCursor;
 	int procSelecter; /* 2:select 3:deciide 4:play 5:result 6:keyconfig 7:skinselect */
 	int procPhase;
-	int hThreadBanner;
+	uintptr_t hThreadBanner;
 	struct gameplay gameplay;
 	CRITICAL_SECTION criticalSection;
 	char is_clicked_screenModeChange;
